% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-patent_scraping.R, R/patent_scraping.R
\name{patent_scraping}
\alias{patent_scraping}
\title{FUNCTION_TITLE}
\usage{
patent_scraping(
  url,
  directory = NULL,
  tables = c("abstract", "claims", "cited_by", "patent_citations", "ipc"),
  show_progress = FALSE
)

patent_scraping(
  url,
  directory = NULL,
  tables = c("abstract", "claims", "cited_by", "patent_citations", "ipc"),
  show_progress = FALSE
)
}
\arguments{
\item{url}{Google Patents URL}

\item{directory}{Path do save patent data (csv's) from google patents, Default: NULL}

\item{tables}{Tables to be extracted, Default: c("abstract", "claims", "cited_by", "patent_citations", "ipc")}

\item{show_progress}{logical, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

Scraping data from google patents
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 # The tables: abstract.csv, claims.csv, cited_by.csv, patent_citations.csv and ipc.csv are 
 # saved in the rawfiles_pantents/ directory.
 patent_scraping(url = 'https://patents.google.com/patent/US10780391B2/en',
                 directory = 'rawfiles_pantents',
                 tables = c('abstract', 'claims', 'cited_by', 'patent_citations', 'ipc'),
                 show_progress = FALSE)
 }
}
}
\seealso{
\code{\link[fs]{create}}
\code{\link[here]{here}}
\code{\link[tibble]{tibble}}
\code{\link[rvest]{reexports}}, \code{\link[rvest]{html_element}}, \code{\link[rvest]{html_text}}, \code{\link[rvest]{html_table}}
\code{\link[readr]{write_delim}}
\code{\link[janitor]{clean_names}}
\code{\link[stats]{setNames}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{relocate}}
\code{\link[stringr]{str_locate}}, \code{\link[stringr]{str_sub}}

\code{\link[fs]{create}}
\code{\link[here]{here}}
\code{\link[tibble]{tibble}}
\code{\link[rvest]{reexports}}, \code{\link[rvest]{html_element}}, \code{\link[rvest]{html_text}}, \code{\link[rvest]{html_table}}
\code{\link[readr]{write_delim}}
\code{\link[janitor]{clean_names}}
\code{\link[stats]{setNames}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{relocate}}
\code{\link[stringr]{str_locate}}, \code{\link[stringr]{str_sub}}
}
