% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-patent_scraping.R
\name{patent_scraping}
\alias{patent_scraping}
\title{FUNCTION_TITLE}
\usage{
patent_scraping(
  url,
  directory = NULL,
  tables = c("abstract", "claims", "cited_by", "patent_citations", "ipc"),
  show_progress = F
)
}
\arguments{
\item{url}{PARAM_DESCRIPTION}

\item{directory}{PARAM_DESCRIPTION, Default: NULL}

\item{tables}{PARAM_DESCRIPTION, Default: c("abstract", "claims", "cited_by", "patent_citations", "ipc")}

\item{show_progress}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fs]{create}}
\code{\link[here]{here}}
\code{\link[tibble]{tibble}}
\code{\link[rvest]{reexports}}, \code{\link[rvest]{html_element}}, \code{\link[rvest]{html_text}}, \code{\link[rvest]{html_table}}
\code{\link[readr]{write_delim}}
\code{\link[janitor]{clean_names}}
\code{\link[stats]{setNames}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{relocate}}
\code{\link[stringr]{str_locate}}, \code{\link[stringr]{str_sub}}
}
